requirements.txt
flask
python-dotenv
requests
aiohttp
google-genai
supabase
dateparser

google-auth
google-auth-oauthlib
google-api-python-client